Reuse main layout for user registration flow
