+++ src/components/frame/v5/pages/CreateUserPage/CreateUserPage.tsx
@@ -9,19 +9,23 @@ const displayName = 'frame.CreateUserPage';
 
 const CreateUserPage = () => {
   return (
-    <ActionForm
-      className="max-w-lg flex flex-col items-end"
-      validationSchema={validationSchema}
-      defaultValues={{
-        username: '',
-        emailAddress: '',
-        emailPermissions: [],
-      }}
-      mode="onChange"
-      actionType={ActionTypes.USERNAME_CREATE}
-    >
-      <CreateUserForm />
-    </ActionForm>
+    <main className="flex flex-col items-center">
+      <article className="max-w-[33.125rem]">
+        <ActionForm
+          className="max-w-lg flex flex-col items-end"
+          validationSchema={validationSchema}
+          defaultValues={{
+            username: '',
+            emailAddress: '',
+            emailPermissions: [],
+          }}
+          mode="onChange"
+          actionType={ActionTypes.USERNAME_CREATE}
+        >
+          <CreateUserForm />
+        </ActionForm>
+      </article>
+    </main>
   );
 };
